requirements.txt
git+https://github.com/openai/CLIP.git
git+https://github.com/facebookresearch/fvcore
accelerate==0.26.1
debugpy==1.6.0
einops==0.6.0
google-auth-oauthlib==0.4.6
huggingface-hub==0.33.1
hydra-core==1.3.2
jedi==0.18.1
numpy==1.22.2
omegaconf==2.3.0
opencv-python==4.7.0.72
packaging==25.0
Pillow==9.4.0
six==1.16.0
stack-data==0.3.0
submitit==1.4.5
tokenizers==0.15.1
tqdm==4.65.0
transformers==4.37.2
typing_extensions==4.5.0
urllib3==1.26.15
wandb==0.14.0
timm==0.4.12
torch_scatter==2.1.2
albumentations==1.3.1
easydict==1.10
volumentations-3D==1.0.4
pycocotools==2.0.7
sentence-transformers==2.2.2
nltk==3.8.1
fpsample==0.3.2
volumentations==0.1.8
charset-normalizer==3.4.2
jsonlines
open3d
hydra
plyfile
volumentations==0.1.8